+++ contracts/gateway/utils/StableCoinFeeRegistry.sol
@@ -10,20 +10,20 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
         string fiatName;
         // The list of erc20 address of stable coin.
         address[] erc20Addresses;
-        // VWBL Fee which is denominated by fiat. 
+        // VWBL Fee which is denominated by fiat.
         // If VWBL Fee is 15 yen, feeNumerator = 15 * 10 ** 10000(_feeDenominator())
-        uint feeNumerator;
+        uint256 feeNumerator;
     }
     // fiatIndex start from 1.
-    uint public nextFiatIndex = 1;
-    mapping (uint => StableCoinInfo) fiatIndexToSCInfo;
-    mapping (address => uint) public erc20ToFiatIndex;
-    uint public registeredTokensCount;
+    uint256 public nextFiatIndex = 1;
+    mapping(uint256 => StableCoinInfo) fiatIndexToSCInfo;
+    mapping(address => uint256) public erc20ToFiatIndex;
+    uint256 public registeredTokensCount;
     address[] public prevRegisteredTokens;
 
     event FeeTokenRegistered(address erc20);
     event FeeTokenUnregistered(address erc20);
-    event StableCoinFeeChanged(string fiatName, uint256 oldFee, uint256 newFee);  
+    event StableCoinFeeChanged(string fiatName, uint256 oldFee, uint256 newFee);
 
     constructor(address _initialOwner) Ownable(_initialOwner) {}
 
@@ -36,9 +36,9 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @return An array of StableCoinInfo structs.
      */
     function getStableCoinInfos() public view returns (StableCoinInfo[] memory) {
-        StableCoinInfo[] memory stableCoinInfos = new StableCoinInfo[](nextFiatIndex-1);
-        uint k = 0;
-        for (uint i = 1; i < nextFiatIndex; i++) {
+        StableCoinInfo[] memory stableCoinInfos = new StableCoinInfo[](nextFiatIndex - 1);
+        uint256 k = 0;
+        for (uint256 i = 1; i < nextFiatIndex; i++) {
             stableCoinInfos[k] = fiatIndexToSCInfo[i];
         }
         return stableCoinInfos;
@@ -50,16 +50,20 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @param _erc20Addresses The list of ERC20 addresses representing the stable coin.
      * @param _feeNumerator The fee numerator denominated in fiat currency.
      */
-    function registerStableCoinInfo(string memory _fiatName, address[] memory _erc20Addresses, uint _feeNumerator) public onlyOwner {
-        for (uint i = 0; i < _erc20Addresses.length; i++) {
+    function registerStableCoinInfo(
+        string memory _fiatName,
+        address[] memory _erc20Addresses,
+        uint256 _feeNumerator
+    ) public onlyOwner {
+        for (uint256 i = 0; i < _erc20Addresses.length; i++) {
             require(!registered(_erc20Addresses[i]), "ERC20 is already registered");
         }
-        uint fiatIndex = nextFiatIndex++;
+        uint256 fiatIndex = nextFiatIndex++;
         StableCoinInfo storage scInfo = fiatIndexToSCInfo[fiatIndex];
         scInfo.fiatName = _fiatName;
         scInfo.erc20Addresses = _erc20Addresses;
         scInfo.feeNumerator = _feeNumerator;
-        for (uint i = 0; i < _erc20Addresses.length; i++) {
+        for (uint256 i = 0; i < _erc20Addresses.length; i++) {
             erc20ToFiatIndex[_erc20Addresses[i]] = fiatIndex;
         }
         registeredTokensCount += _erc20Addresses.length;
@@ -70,7 +74,7 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @param fiatIndex The index of the fiat currency to be renamed.
      * @param newFiatName The new name for the fiat currency.
      */
-    function renameFiat(uint fiatIndex, string memory newFiatName) public onlyOwner {
+    function renameFiat(uint256 fiatIndex, string memory newFiatName) public onlyOwner {
         require(fiatIndex < nextFiatIndex, "fiatIndex is invalid");
         fiatIndexToSCInfo[fiatIndex].fiatName = newFiatName;
     }
@@ -80,15 +84,15 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @param fiatIndex The index of the stable coin to register the ERC20 addresses for.
      * @param newERC20Addresses The list of new ERC20 addresses to register.
      */
-    function registerERC20Addresses(uint fiatIndex, address[] memory newERC20Addresses) public onlyOwner {
+    function registerERC20Addresses(uint256 fiatIndex, address[] memory newERC20Addresses) public onlyOwner {
         require(fiatIndex < nextFiatIndex, "fiatIndex is invalid");
         StableCoinInfo storage scInfo = fiatIndexToSCInfo[fiatIndex];
-        for (uint i = 0; i < newERC20Addresses.length; i++) {
+        for (uint256 i = 0; i < newERC20Addresses.length; i++) {
             require(!registered(newERC20Addresses[i]), "This ERC20 is already registered");
         }
 
         registeredTokensCount += newERC20Addresses.length;
-        for (uint i = 0; i < newERC20Addresses.length; i++) {
+        for (uint256 i = 0; i < newERC20Addresses.length; i++) {
             scInfo.erc20Addresses.push(newERC20Addresses[i]);
             erc20ToFiatIndex[newERC20Addresses[i]] = fiatIndex;
             emit FeeTokenRegistered(newERC20Addresses[i]);
@@ -101,15 +105,15 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @param fiatIndex The index of the stable coin to unregister the ERC20 address from.
      * @param erc20Address The address of the ERC20 token to unregister.
      */
-    function unregisterERC20Address(uint fiatIndex, address erc20Address) public onlyOwner {
+    function unregisterERC20Address(uint256 fiatIndex, address erc20Address) public onlyOwner {
         require(fiatIndex < nextFiatIndex, "fiatIndex is invalid");
         require(registered(erc20Address), "This ERC20 is not registered");
 
         registeredTokensCount -= 1;
         StableCoinInfo storage scInfo = fiatIndexToSCInfo[fiatIndex];
-        address[] memory newERC20Addresses = new address[](scInfo.erc20Addresses.length-1);
-        uint j = 0;
-        for (uint i = 0; i < scInfo.erc20Addresses.length; i++) {
+        address[] memory newERC20Addresses = new address[](scInfo.erc20Addresses.length - 1);
+        uint256 j = 0;
+        for (uint256 i = 0; i < scInfo.erc20Addresses.length; i++) {
             if (scInfo.erc20Addresses[i] != erc20Address) {
                 newERC20Addresses[j] = scInfo.erc20Addresses[i];
                 j++;
@@ -119,7 +123,7 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
         delete erc20ToFiatIndex[erc20Address];
 
         bool prevRegistered = false;
-        for (uint i = 0; i < prevRegisteredTokens.length; i++) {
+        for (uint256 i = 0; i < prevRegisteredTokens.length; i++) {
             if (prevRegisteredTokens[i] == erc20Address) {
                 prevRegistered = true;
             }
@@ -136,10 +140,10 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @param fiatIndex The index of the stable coin to register the fee numerator for.
      * @param newFeeNumerator The new fee numerator denominated in fiat currency to register.
      */
-    function registerFeeNumerator(uint fiatIndex, uint newFeeNumerator) public onlyOwner {
+    function registerFeeNumerator(uint256 fiatIndex, uint256 newFeeNumerator) public onlyOwner {
         require(fiatIndex < nextFiatIndex, "fiatIndex is invalid");
         StableCoinInfo storage scInfo = fiatIndexToSCInfo[fiatIndex];
-        uint oldFeeNumerator = scInfo.feeNumerator;
+        uint256 oldFeeNumerator = scInfo.feeNumerator;
         require(oldFeeNumerator != newFeeNumerator);
         scInfo.feeNumerator = newFeeNumerator;
         emit StableCoinFeeChanged(scInfo.fiatName, oldFeeNumerator, newFeeNumerator);
@@ -151,7 +155,9 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @return bool Returns true if the ERC20 token is registered, false otherwise.
      */
     function registered(address erc20Address) public view returns (bool) {
-        if (erc20ToFiatIndex[erc20Address] != 0) {return true;}
+        if (erc20ToFiatIndex[erc20Address] != 0) {
+            return true;
+        }
         return false;
     }
 
@@ -165,10 +171,10 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
         if (!registered(erc20Address)) {
             return (0, false);
         }
-        uint fiatIndex = erc20ToFiatIndex[erc20Address];
-        uint feeNumerator = fiatIndexToSCInfo[fiatIndex].feeNumerator;
+        uint256 fiatIndex = erc20ToFiatIndex[erc20Address];
+        uint256 feeNumerator = fiatIndexToSCInfo[fiatIndex].feeNumerator;
         uint8 decimals = ERC20(erc20Address).decimals();
-        uint feeDecimals = feeNumerator * (10**decimals / _feeDenominator());
+        uint256 feeDecimals = feeNumerator * (10**decimals / _feeDenominator());
         return (feeDecimals, true);
     }
 
@@ -178,9 +184,9 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      */
     function getRegisteredTokens() public view returns (address[] memory) {
         address[] memory registeredTokens = new address[](registeredTokensCount);
-        uint k = 0;
-        for (uint i = 1; i < nextFiatIndex; i++) {
-            for (uint j = 0; j < fiatIndexToSCInfo[i].erc20Addresses.length; j++) {
+        uint256 k = 0;
+        for (uint256 i = 1; i < nextFiatIndex; i++) {
+            for (uint256 j = 0; j < fiatIndexToSCInfo[i].erc20Addresses.length; j++) {
                 registeredTokens[k] = fiatIndexToSCInfo[i].erc20Addresses[j];
                 k += 1;
             }
@@ -192,7 +198,7 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @notice Returns the total number of registered ERC20 Fee tokens.
      * @return The total count of registered ERC20 Fee tokens.
      */
-    function getRegisteredTokensCount() public view returns (uint) {
+    function getRegisteredTokensCount() public view returns (uint256) {
         return registeredTokensCount;
     }
 
@@ -208,7 +214,7 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @notice Returns the total number of previously registered ERC20 Fee tokens.
      * @return The total count of previous registered ERC20 Fee tokens.
      */
-    function getPrevRegisteredTokensCount() public view returns (uint) {
+    function getPrevRegisteredTokensCount() public view returns (uint256) {
         return prevRegisteredTokens.length;
     }
 
@@ -217,11 +223,11 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @return An array of addresses the previously and currently registered ERC20 Fee tokens.
      */
     function getPrevAndCurRegisteredTokens() public view returns (address[] memory) {
-        uint resultTokensCount = getRegisteredTokensCount() + getPrevRegisteredTokensCount();
+        uint256 resultTokensCount = getRegisteredTokensCount() + getPrevRegisteredTokensCount();
         address[] memory resultTokens = new address[](resultTokensCount);
         resultTokens = getRegisteredTokens();
-        uint j = 0;
-        for (uint i = getRegisteredTokensCount(); i < resultTokensCount; i++) {
+        uint256 j = 0;
+        for (uint256 i = getRegisteredTokensCount(); i < resultTokensCount; i++) {
             resultTokens[i] = prevRegisteredTokens[j];
             j += 1;
         }
@@ -232,7 +238,7 @@ contract StableCoinFeeRegistry is IStableCoinFeeRegistry, Ownable {
      * @notice Returns the total number of previously and currently registered ERC20 Fee tokens in the fee registry.
      * @return The total count of previously and currently registered ERC20 Fee tokens.
      */
-    function getPrevAndCurRegisteredTokensCount() public view returns (uint) {
+    function getPrevAndCurRegisteredTokensCount() public view returns (uint256) {
         return registeredTokensCount + prevRegisteredTokens.length;
     }
 }
