+++ contracts/gateway/utils/WithdrawExtraFee.sol
@@ -6,14 +6,12 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "./IWithdrawExtraFee.sol";
 
 contract WithdrawExtraFee is Ownable, ReentrancyGuard, IWithdrawExtraFee {
-    uint public totalPendingExtraFee;
-    mapping (address => uint) public srcAddrToPendingExtraFee;
-    mapping (address => address) public srcAddrToRecipient;
+    uint256 public totalPendingExtraFee;
+    mapping(address => uint256) public srcAddrToPendingExtraFee;
+    mapping(address => address) public srcAddrToRecipient;
     address[] public srcAddresses;
 
-    constructor(
-        address _initialOwner
-    ) Ownable(_initialOwner) {}
+    constructor(address _initialOwner) Ownable(_initialOwner) {}
 
     /**
      * @notice Deposit extra vwbl fee into the contract from specified source address.
@@ -32,8 +30,11 @@ contract WithdrawExtraFee is Ownable, ReentrancyGuard, IWithdrawExtraFee {
      * @param srcAddress The address who call payFee method of VWBLGatewayV2 contract.
      */
     function withdrawAllPendingToken(address srcAddress) public nonReentrant {
-        require(srcAddress == msg.sender || srcAddrToRecipient[srcAddress] == msg.sender, "msg sender is invalid withdrawer");
-        uint pendingAmount = srcAddrToPendingExtraFee[srcAddress];
+        require(
+            srcAddress == msg.sender || srcAddrToRecipient[srcAddress] == msg.sender,
+            "msg sender is invalid withdrawer"
+        );
+        uint256 pendingAmount = srcAddrToPendingExtraFee[srcAddress];
         require(pendingAmount > 0, "pending extra fee amount is 0");
 
         address recipient;
@@ -52,9 +53,12 @@ contract WithdrawExtraFee is Ownable, ReentrancyGuard, IWithdrawExtraFee {
      * @notice Withdraws pending extra fees for a specified source address.
      * @param srcAddress The address who call payFee method of VWBLGatewayV2 contract.
      */
-    function withdrawPendingToken(address srcAddress, uint withdrawAmount) public nonReentrant {
-        require(srcAddress == msg.sender || srcAddrToRecipient[srcAddress] == msg.sender, "msg sender is invalid withdrawer");
-        uint pendingAmount = srcAddrToPendingExtraFee[srcAddress];
+    function withdrawPendingToken(address srcAddress, uint256 withdrawAmount) public nonReentrant {
+        require(
+            srcAddress == msg.sender || srcAddrToRecipient[srcAddress] == msg.sender,
+            "msg sender is invalid withdrawer"
+        );
+        uint256 pendingAmount = srcAddrToPendingExtraFee[srcAddress];
         require(withdrawAmount <= pendingAmount, "withdraw amount is insufficent than pending amount");
 
         address recipient;
@@ -85,9 +89,9 @@ contract WithdrawExtraFee is Ownable, ReentrancyGuard, IWithdrawExtraFee {
      * @param rmSrcAddress The address to be removed from the list of source addresses.
      */
     function rmFromSrcAddresses(address rmSrcAddress) private {
-        address[] memory newSrcAddresses = new address[](srcAddresses.length-1);
-        uint j = 0;
-        for (uint i = 0; i < srcAddresses.length; i++) {
+        address[] memory newSrcAddresses = new address[](srcAddresses.length - 1);
+        uint256 j = 0;
+        for (uint256 i = 0; i < srcAddresses.length; i++) {
             if (rmSrcAddress != srcAddresses[i]) {
                 newSrcAddresses[j] = srcAddresses[i];
             }
